fix(db): close invalid states, races, and type gaps found in the schema review
Migration 253 hardens the schema: shape CHECKs (card_bans date order,
copy-entry quantity, pending-trade expiry, job_runs status shapes, card
token self-reference, meta source all-or-none, lookup slug/label not
empty), ON DELETE CASCADE for marketplace_ignored_variants so clearing
price data works with ignored variants, provider-scoped printing link
overrides, tournament-scoped team FKs, unique running-job and pending
Discord-code indexes, FK indexes for printing_images and
candidate_meta_decks, and prefix-redundant index cleanup.

Migration 254 drops organizations.owner_user_id: ownership is the
role = 'owner' membership rows alone, guarded by a deferred trigger
that keeps every org at one owner or more; the user-deletion rebalance
keys on the role.

Repository fixes: transactional junction replaces and pending-link
creation, copy locks before bulk-dispose pin checks, locked re-checks
for manual deck-check entries, participant removal, player deck-check
transitions and deck-family repairs, an atomic trade-channel toggle,
job-claim via the unique index, overlay-channel race recovery, a
teamId-null guard on team creation, surfaced gallery-accept skips, and
honest claim-URL reporting.

Type-safety: marketplace columns carry the Marketplace union end to
end, and aggregate sql casts match what the driver returns. Integration
fixtures that violated the migration 245-252 constraints are repaired.

+++ apps/api/src/repositories/status.ts
@@ -49,8 +49,9 @@ export function statusRepo(db: Kysely<Database>) {
     async getDatabaseStatus(): Promise<DbStatus> {
       try {
         // Database size
+        // ::float8 because the numeric division would come back as a string.
         const [sizeRow] = await sql<{ sizeMb: number }>`
-          SELECT pg_database_size(current_database()) / (1024 * 1024.0) AS size_mb
+          SELECT (pg_database_size(current_database()) / (1024 * 1024.0))::float8 AS size_mb
         `
           .execute(db)
           .then((r) => r.rows);
